Abstracted border color class assembly into a method
Only apply border color to the input and button elements when the button is outside.
When the button is inside, we apply the border color styles to the outer wrapper.

Declaring variables for button positions to avoid repetition
Applying `borderProps.style` to the text field and button when the button is not inside so we can ensure that all future supports styles such as border width and style are applied to the element.